Decode data to binary, not atoms
Decoding to atoms is bad because atoms are not garbage collected. We have a case where data in our ES stores can contain user input data, in a structure like this:

{
  total_comments: 1,
  topics: {
    topic: 1,
    another_topic: 2
  }
}

Previously, tirexs would decode the "topic" and "another_topic" keys to atoms, rather than strings. Doing so may lead to an attack vector that people may exploit.